modularize helper functions into lib
